split out docker/rkt structs from the framework
